+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.6.13
+
+### Patch Changes
+
+- c707af1: fix(deps): update dependency dotenv to v16.6.1
+
 ## 1.6.12
 
 ### Patch Changes